Adding discret reward scheme an shift of reward value
Instead of using the idle time of vertices, it is also possible to use a rank of the idle time of each node, ranging from 0 to n, where n is the number of nodes. Discrete linear ranking seems to make learning easier initially. There's also a shift in the reward.

If the shift is set to 0.3, it means that the nodes visited most recently (top 30%) will not yield any reward when visited.